+++ src/Spy/Spy.php
@@ -306,6 +306,36 @@ class Spy extends AbstractWrappedInvocable implements SpyInterface
     }
 
     /**
+     * Get the first event.
+     *
+     * @return EventInterface          The event.
+     * @throws UndefinedEventException If there are no events.
+     */
+    public function firstEvent()
+    {
+        if (!$this->calls) {
+            throw new UndefinedEventException(0);
+        }
+
+        return $this->calls[0];
+    }
+
+    /**
+     * Get the last event.
+     *
+     * @return EventInterface          The event.
+     * @throws UndefinedEventException If there are no events.
+     */
+    public function lastEvent()
+    {
+        if ($count = count($this->calls)) {
+            return $this->calls[$count - 1];
+        }
+
+        throw new UndefinedEventException(0);
+    }
+
+    /**
      * Get an event by index.
      *
      * Negative indices are offset from the end of the list. That is, `-1`